ATST-SED/requirements.txt
dcase_util==0.2.20
einops==0.8.0
h5py==3.11.0
librosa==0.10.2.post1
matplotlib==3.7.2
numpy==1.23.5
pandas==2.0.3
psds_eval==0.5.3
pytorch_lightning==2.4.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.1
sed_eval==0.2.1
sed_scores_eval==0.0.0
setuptools==75.2.0
soundfile==0.12.1
thop==0.1.1.post2209072238
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.5.2
tqdm==4.66.5
